+++ covalent_dispatcher/_cli/service.py
@@ -305,6 +305,7 @@ def start(
         {
             "user_interface.address": "0.0.0.0",
             "user_interface.port": port,
+            "develop": develop,
             "dispatcher.address": "0.0.0.0",
             "dispatcher.port": port,
             "dask": {
@@ -350,6 +351,7 @@ def restart(ctx, port: bool, develop: bool) -> None:
     """
     Restart the server.
     """
+    set_config({"develop": develop})
     port = port or get_config("user_interface.port")
 
     ctx.invoke(stop)