+++ covalent_dispatcher/_core/execution.py
@@ -276,7 +276,7 @@ def _run_planned_workflow(result_object: Result, thread_pool: ThreadPoolExecutor
         result_object._update_node(**node_result)
         with DispatchDB() as db:
             db.upsert(result_object.dispatch_id, result_object)
-        result_object.save()
+            db.save_db(result_object)
         result_webhook.send_update(result_object)
 
     def task_callback(future: Future):
@@ -397,7 +397,7 @@ def _run_planned_workflow(result_object: Result, thread_pool: ThreadPoolExecutor
                 result_object._error = f"Node {result_object._get_node_name(node_id)} failed: \n{result_object._get_node_error(node_id)}"
                 with DispatchDB() as db:
                     db.upsert(result_object.dispatch_id, result_object)
-                result_object.save()
+                    db.save_db(result_object)
                 result_webhook.send_update(result_object)
                 return
 
@@ -406,7 +406,7 @@ def _run_planned_workflow(result_object: Result, thread_pool: ThreadPoolExecutor
                 result_object._end_time = datetime.now(timezone.utc)
                 with DispatchDB() as db:
                     db.upsert(result_object.dispatch_id, result_object)
-                result_object.save()
+                    db.save_db(result_object)
                 result_webhook.send_update(result_object)
                 return
 
@@ -419,7 +419,7 @@ def _run_planned_workflow(result_object: Result, thread_pool: ThreadPoolExecutor
     result_object._end_time = datetime.now(timezone.utc)
     with DispatchDB() as db:
         db.upsert(result_object.dispatch_id, result_object)
-    result_object.save(write_source=True)
+        db.save_db(result_object, write_source=True)
     result_webhook.send_update(result_object)
 
 
@@ -473,7 +473,7 @@ def run_workflow(dispatch_id: str, results_dir: str, tasks_pool: ThreadPoolExecu
         result_object._status = Result.FAILED
         result_object._end_time = datetime.now(timezone.utc)
         result_object._error = "".join(traceback.TracebackException.from_exception(ex).format())
-        result_object.save()
+        DispatchDB().save_db(result_object)
         raise