+++ covalent_dispatcher/_db/dispatchdb.py
@@ -8,11 +8,15 @@ import networkx as nx
 import simplejson
 
 import covalent.executor as covalent_executor
+from covalent._data_store import DataStore
 from covalent._results_manager.result import Result
+from covalent._shared_files import logger
 from covalent._shared_files.config import get_config
 from covalent._shared_files.util_classes import Status
 from covalent._shared_files.utils import get_named_params
 
+app_log = logger.app_log
+log_stack_info = logger.log_stack_info
 # DB Schema:
 # TABLE dispatches
 # * dispatch_id text primary key
@@ -166,6 +170,23 @@ class DispatchDB:
         self.conn.close()
 
         return False
+
+    def _db_dev_path(self):
+        if get_config("develop"):
+            sqlite = ".sqlite"
+            path = "sqlite+pysqlite:///" + self._dbpath.split(sqlite)[0] + "_dev" + sqlite
+            print("***SET DB PATH FOR DEVELOPMENT***")
+            print(path)
+            return path
+        else:
+            return None
+
+    def save_db(self, result_object: Result, **kwargs):
+        try:
+            result_object.save(**kwargs)
+            result_object.persist(DataStore(self._db_dev_path()))
+        except Exception:
+            app_log.exception("Exception occured while saving to DB.")
 
     def get(self, dispatch_ids: [] = []) -> List[Tuple[str, str]]:
         """